+++ Assets/GraphicsLabor/Scripts/Attributes/LaborerAttributes/InspectedAttributes/LabelAttribute.cs
@@ -1,10 +1,10 @@
 ﻿namespace GraphicsLabor.Scripts.Attributes.LaborerAttributes.InspectedAttributes
 {
-    public abstract class LabelAttribute: InspectedAttribute
+    public class LabelAttribute: InspectedAttribute
     {
         public readonly string _label;
 
-        protected LabelAttribute(string label)
+        public LabelAttribute(string label)
         {
             _label = label;
         }