+++ Assets/GraphicsLabor/Scripts/Core/LaborerTags/LaborTagComponent.cs
@@ -1,0 +1,7 @@
+﻿namespace GraphicsLabor.Scripts.Core.LaborerTags
+{
+    public class LaborTagComponent
+    {
+        
+    }
+}