+++ Assets/GraphicsLabor/Scripts/Editor/Windows/EditorWindowBase.cs
@@ -1,5 +1,9 @@
 using System;
+using GraphicsLabor.Scripts.Core.Settings;
+using GraphicsLabor.Scripts.Core.Utility;
 using GraphicsLabor.Scripts.Editor.Settings;
+using UnityEditor.Callbacks;
+using UnityEngine;
 using Object = UnityEngine.Object;
 
 
@@ -8,7 +12,6 @@ namespace GraphicsLabor.Scripts.Editor.Windows
     public abstract class EditorWindowBase : WindowBase
     {
         protected abstract void PassInspectedObject(Object obj);
-
         
         protected static T CreateNewEditorWindow<T>(Object obj, string displayName = "EditorWindowBase") where T : EditorWindowBase
         {
@@ -18,7 +21,7 @@ namespace GraphicsLabor.Scripts.Editor.Windows
             if (settings.OpenedCustomWindows.Count != 0)
             {
                 window = settings.FindWindowWhere(customEditor =>
-                    customEditor.WindowName == (obj != null ? obj.name : "null") && customEditor.SelfType == typeof(T));
+                    (customEditor.WindowName == (obj != null ? obj.name : "null") || customEditor.WindowName == displayName) && customEditor.SelfType == typeof(T));
             }
             
             if (window == null)
@@ -27,6 +30,7 @@ namespace GraphicsLabor.Scripts.Editor.Windows
             }
             
             window.Focus();
+            (window as T)?.PassInspectedObject(obj);
             return window as T;
         }
 
@@ -35,6 +39,23 @@ namespace GraphicsLabor.Scripts.Editor.Windows
             T window = WindowBase.CreateAndInitWindow<T>(displayName, desiredDockNextTo);
             window.PassInspectedObject(obj);
             return window;
+        }
+
+        [DidReloadScripts]
+        private static void OnScriptReloadSelf()
+        {
+            GLogger.Log("Reloading scripts");
+            
+            WindowSettings settings = GetWindowSettings();
+            if (settings.OpenedCustomWindows.Count == 0)
+            {
+                settings.OpenedCustomWindows.AddRange(Resources.FindObjectsOfTypeAll<WindowBase>());
+
+                foreach (WindowBase windowBase in settings.OpenedCustomWindows)
+                {
+                    windowBase.SetType(windowBase.GetType());
+                }
+            }
         }